Title: 4가지 연산을 이용하여 1 만들기, Time: 1273ms, Memory: 176MB, Status: Memory Limit Exceed - CodeTree

+++ 241023/4가지 연산을 이용하여 1 만들기/make-one-using-four-operations.java
@@ -5,7 +5,7 @@ public class Main {
     static StringTokenizer tokens;
     static BufferedReader input = new BufferedReader(new InputStreamReader(System.in));
     static StringBuilder output = new StringBuilder();
-    static int n,res;
+    static long n,res;
     public static void main(String[] args) throws IOException {
         n = Integer.parseInt(input.readLine());
 
@@ -15,21 +15,21 @@ public class Main {
 
     }
     static void bfs() {
-        Queue<int []> que = new ArrayDeque<>();
-        que.offer(new int[] {n, 0});
+        Queue<long []> que = new ArrayDeque<>();
+        que.offer(new long[] {n, 0});
         while(!que.isEmpty()) {
-            int t[] = que.poll();
-            int x = t[0];
-            int cnt = t[1];
+            long t[] = que.poll();
+            long x = t[0];
+            long cnt = t[1];
             //System.out.println(x);
             if(x == 1){
                 res = cnt;
                 break;
             }
-            que.offer(new int[] {x+1, cnt+1});
-            que.offer(new int[] {x-1, cnt+1});
-            if(x%2==0) que.offer(new int[] {x/2, cnt+1});
-            if(x%3==0) que.offer(new int[] {x/3, cnt+1});
+            que.offer(new long[] {x+1, cnt+1});
+            que.offer(new long[] {x-1, cnt+1});
+            if(x%2==0) que.offer(new long[] {x/2, cnt+1});
+            if(x%3==0) que.offer(new long[] {x/3, cnt+1});
         }
 
     }
